抽离常量文件
